+++ assets/src/components/activities/multiple_choice/transformations/v2.ts
@@ -1,37 +1,21 @@
+import { DEFAULT_PART_ID } from 'components/activities/common/utils';
+import { MCSchemaV1 } from 'components/activities/multiple_choice/transformations/v1';
 import {
   ActivityModelSchema,
   Choice,
   ChoiceIdsToResponseId,
-  makeResponse,
   Part,
   Stem,
   Transformation,
 } from 'components/activities/types';
-import {
-  getCorrectResponse,
-  getResponses,
-} from 'components/activities/common/responses/authoring/responseUtils';
-import { matchRule } from 'components/activities/common/responses/authoring/rules';
-import { MCSchemaV1 } from 'components/activities/multiple_choice/transformations/v1';
+import { getCorrectResponse, getResponses, Responses } from 'data/activities/model/responses';
+import { matchRule } from 'data/activities/model/rules';
 
 export interface MCSchemaV2 extends ActivityModelSchema {
   stem: Stem;
   choices: Choice[];
   authoring: {
     version: 2;
-    targeted: ChoiceIdsToResponseId[];
-    parts: Part[];
-    transformations: Transformation[];
-    previewText: string;
-  };
-}
-
-export interface CATASchemaV2 extends ActivityModelSchema {
-  stem: Stem;
-  choices: Choice[];
-  authoring: {
-    version: 2;
-    // An association list of correct choice ids to the matching response id
     targeted: ChoiceIdsToResponseId[];
     parts: Part[];
     transformations: Transformation[];
@@ -54,8 +38,8 @@ export const mcV1toV2 = (model: MCSchemaV1): MCSchemaV2 => {
 
   if (!getResponses(newModel).find((r) => r.rule === matchRule('.*'))) {
     newModel.authoring.parts[0].responses = [
-      getCorrectResponse(newModel),
-      makeResponse(matchRule('.*'), 0, ''),
+      getCorrectResponse(newModel, DEFAULT_PART_ID),
+      Responses.catchAll(),
     ];
   }