+++ assets/src/data/content/writers/context.ts
@@ -1,5 +1,17 @@
+import { MultiInputDelivery } from 'components/activities/multi_input/schema';
+import { ID } from 'data/content/model';
+
 export interface WriterContext {
   sectionSlug?: string;
+  inputRefContext?: {
+    onChange: (id: string, e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => void;
+    toggleHints: (id: string) => void;
+    inputs: Map<
+      ID,
+      { input: MultiInputDelivery; value: string; placeholder?: string; hasHints: boolean }
+    >;
+    disabled: boolean;
+  };
 }
 
 export const defaultWriterContext = (params: Partial<WriterContext> = {}) =>