+++ lib/oli_web/live/delivery/manage_section.ex
@@ -3,11 +3,10 @@ defmodule OliWeb.Delivery.ManageSection do
 
   import OliWeb.ViewHelpers,
     only: [
-      is_admin?: 1,
-      user_role: 2
+      is_section_instructor_or_admin?: 2,
+      is_admin?: 2
     ]
 
-  alias Lti_1p3.Tool.ContextRoles
   alias Oli.Delivery.Sections
   alias OliWeb.Router.Helpers, as: Routes
   alias Oli.Accounts
@@ -18,16 +17,12 @@ defmodule OliWeb.Delivery.ManageSection do
         %{"section_slug" => section_slug, "current_user_id" => current_user_id},
         socket
       ) do
-    section = Sections.get_section_by_slug(section_slug)
     current_user = Accounts.get_user!(current_user_id) |> Repo.preload([:platform_roles, :author])
 
-    # only permit instructor level access
-    if is_admin?(%{assigns: %{current_author: current_user.author}}) or
-         ContextRoles.has_role?(
-           current_user,
-           section.slug,
-           ContextRoles.get_role(:context_instructor)
-         ) do
+    # only permit instructor or admin level access
+    if is_section_instructor_or_admin?(section_slug, current_user) do
+      section = Sections.get_section_by_slug(section_slug)
+
       socket =
         socket
         |> assign(:section, section)
@@ -51,7 +46,7 @@ defmodule OliWeb.Delivery.ManageSection do
 
       <h2><%= dgettext("section", "Manage Section") %></h2>
 
-      <%= if user_role(@section, @current_user) == :administrator do %>
+      <%= if is_admin?(@section.slug, @current_user) do %>
         <div class="card border-warning my-4">
           <h6 class="card-header">
             Admin Tools