+++ lib/oli_web/views/view_helpers.ex
@@ -6,6 +6,7 @@ defmodule OliWeb.ViewHelpers do
 
   alias Lti_1p3.Tool.ContextRoles
   alias Lti_1p3.Tool.PlatformRoles
+  alias Oli.Delivery.Sections
   alias Oli.Delivery.Sections.Section
   alias Oli.Branding.Brand
 
@@ -30,11 +31,6 @@ defmodule OliWeb.ViewHelpers do
     """
   end
 
-  def is_admin?(%{:assigns => assigns}) do
-    admin_role_id = Oli.Accounts.SystemRole.role_id().admin
-    assigns.current_author.system_role_id == admin_role_id
-  end
-
   def preview_mode(%{assigns: assigns} = _conn) do
     Map.get(assigns, :preview_mode, false)
   end
@@ -48,64 +44,29 @@ defmodule OliWeb.ViewHelpers do
     end
   end
 
-  @admin_roles [
-    PlatformRoles.get_role(:system_administrator),
-    PlatformRoles.get_role(:institution_administrator),
-    ContextRoles.get_role(:context_administrator)
-  ]
+  def is_section_instructor_or_admin?(section_slug, user) do
+    is_section_instructor?(section_slug, user) || is_admin?(section_slug, user)
+  end
 
-  @instructor_roles [
-    PlatformRoles.get_role(:institution_instructor),
-    ContextRoles.get_role(:context_instructor)
-  ]
+  def is_section_instructor?(section_slug, user) do
+    Sections.is_enrolled?(user.id, section_slug) &&
+      ContextRoles.has_role?(
+        user,
+        section_slug,
+        ContextRoles.get_role(:context_instructor)
+      )
+  end
 
-  @student_roles [
-    PlatformRoles.get_role(:institution_student),
-    PlatformRoles.get_role(:institution_learner),
-    ContextRoles.get_role(:context_learner)
-  ]
-
-  def user_role(section, user) do
-    case section do
-      %Section{open_and_free: open_and_free, slug: section_slug} ->
-        cond do
-          open_and_free ->
-            :open_and_free
-
-          PlatformRoles.has_roles?(user, @admin_roles, :any) ||
-              ContextRoles.has_roles?(user, section_slug, @admin_roles, :any) ->
-            :administrator
-
-          PlatformRoles.has_roles?(user, @instructor_roles, :any) ||
-              ContextRoles.has_roles?(user, section_slug, @instructor_roles, :any) ->
-            :instructor
-
-          PlatformRoles.has_roles?(user, @student_roles, :any) ||
-              ContextRoles.has_roles?(user, section_slug, @student_roles, :any) ->
-            :student
-
-          true ->
-            :other
-        end
-
-      _ ->
-        cond do
-          user.guest ->
-            :open_and_free
-
-          PlatformRoles.has_roles?(user, @admin_roles, :any) ->
-            :administrator
-
-          PlatformRoles.has_roles?(user, @instructor_roles, :any) ->
-            :instructor
-
-          PlatformRoles.has_roles?(user, @student_roles, :any) ->
-            :student
-
-          true ->
-            :other
-        end
-    end
+  def is_admin?(section_slug, user) do
+    PlatformRoles.has_roles?(
+      user,
+      [
+        PlatformRoles.get_role(:system_administrator),
+        PlatformRoles.get_role(:institution_administrator)
+      ],
+      :any
+    ) ||
+      ContextRoles.has_role?(user, section_slug, ContextRoles.get_role(:context_administrator))
   end
 
   def maybe_section_slug(conn) do